# contracts/M05/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "M05 Assessments, Grades and Published Answer Sheets",
    "version": "1.0.0",
    "description": "Paths relative to /api/v1. Closed write schemas. Publication requires Idempotency-Key and recent 2FA. Draft results never returned to student/guardian scopes."
  },
  "servers": [
    {
      "url": "/api/v1"
    }
  ],
  "security": [
    {
      "session": []
    }
  ],
  "paths": {
    "/assessments": {
      "post": {
        "operationId": "create_assessment",
        "summary": "Create assessment structure",
        "x-permission": "assessment.manage",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateAssessmentRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AssessmentDTO"
                }
              }
            }
          },
          "401": {
            "description": "error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "403": {
            "description": "error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "422": {
            "description": "error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        }
      },
      "get": {
        "operationId": "list_assessments",
        "summary": "Assessments for the caller's classes (all for school-wide readers)",
        "responses": {
          "200": {
            "description": "Newest first.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "items": {
                      "type": "array",
                      "items": {
                        "type": "object"
                      }
                    },
                    "next_cursor": {
                      "type": [
                        "string",
                        "null"
                      ]
                    }
                  },
                  "required": [
                    "items",
                    "next_cursor"
                  ]
                }
              }
            }
          }
        }
      }
    },
    "/assessments/{id}/results/{student_id}": {
      "patch": {
        "operationId": "patch_result_marks",
        "summary": "Edit marks for one student result",
        "x-permission": "marks.edit",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "student_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PatchResultRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PatchResultResponse"
                }
              }
            }
          },
          "401": {
            "description": "error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "403": {
            "description": "error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "404": {
            "description": "error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "409": {
            "description": "error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "422": {
            "description": "error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        }
      }
    },
    "/results/{id}/evidence": {
      "post": {
        "operationId": "bind_result_evidence",
        "summary": "Bind ordered quality-confirmed answer-sheet pages",
        "x-permission": "marks.edit",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BindEvidenceRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Bound",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BindEvidenceResponse"
                }
              }
            }
          },
          "401": {
            "description": "error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "403": {
            "description": "error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "404": {
            "description": "error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "409": {
            "description": "error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "422": {
            "description": "error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        }
      }
    },
    "/assessments/{id}/submit": {
      "post": {
        "operationId": "submit_assessment",
        "summary": "Submit assessment for review",
        "x-permission": "marks.submit",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ExpectedVersionBody"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Submitted",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AssessmentDTO"
                }
              }
            }
          },
          "401": {
            "description": "error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "403": {
            "description": "error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "404": {
            "description": "error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "409": {
            "description": "error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "422": {
            "description": "error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        }
      }
    },
    "/assessments/{id}/approve": {
      "post": {
        "operationId": "approve_assessment",
        "summary": "Approve submitted assessment",
        "x-permission": "results.approve",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ExpectedVersionBody"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Approved",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AssessmentDTO"
                }
              }
            }
          },
          "401": {
            "description": "error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "403": {
            "description": "error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "404": {
            "description": "error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "409": {
            "description": "error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "422": {
            "description": "error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        }
      }
    },
    "/assessments/{id}/publication": {
      "post": {
        "operationId": "publish_assessment",
        "summary": "Publish approved results; Idempotency-Key required",
        "x-permission": "results.publish",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PublicationRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Published",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicationResponse"
                }
              }
            }
          },
          "401": {
            "description": "error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "403": {
            "description": "error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "404": {
            "description": "error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "409": {
            "description": "error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "422": {
            "description": "error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        }
      }
    },
    "/assessments/{id}/reopen": {
      "post": {
        "operationId": "reopen_assessment",
        "summary": "Reopen published assessment; reason required; recent 2FA",
        "x-permission": "results.reopen",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ExpectedVersionBody"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Reopened",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AssessmentDTO"
                }
              }
            }
          },
          "401": {
            "description": "error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "403": {
            "description": "error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "404": {
            "description": "error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "409": {
            "description": "error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "422": {
            "description": "error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "ErrorEnvelope": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "code",
          "message_key",
          "request_id",
          "field_errors"
        ],
        "properties": {
          "code": {
            "type": "string",
            "enum": [
              "unauthenticated",
              "stale_auth",
              "action_denied",
              "object_inaccessible",
              "version_conflict",
              "state_conflict",
              "validation_failed",
              "rate_limited"
            ]
          },
          "message_key": {
            "type": "string",
            "minLength": 1
          },
          "request_id": {
            "type": "string",
            "minLength": 1
          },
          "field_errors": {
            "type": "array",
            "items": {
              "type": "object",
              "additionalProperties": false,
              "required": [
                "field",
                "message_key"
              ],
              "properties": {
                "field": {
                  "type": "string"
                },
                "message_key": {
                  "type": "string"
                }
              }
            }
          }
        }
      },
...
      "CreateAssessmentRequest": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "year_id",
          "term_id",
          "section_id",
          "subject_id",
          "type",
          "components",
          "policy_version"
        ],
        "properties": {
          "year_id": {
            "type": "string",
            "format": "uuid"
          },
          "term_id": {
            "type": "string",
            "format": "uuid"
          },
          "section_id": {
            "type": "string",
            "format": "uuid"
          },
          "subject_id": {
            "type": "string",
            "format": "uuid"
          },
          "type": {
            "$ref": "#/$defs/AssessmentType"
          },
          "components": {
            "type": "array",
            "minItems": 1,
            "items": {
              "$ref": "#/$defs/ComponentInput"
            }
          },
          "policy_version": {
            "type": "string",
            "minLength": 1,
            "maxLength": 64
          },
          "due_at": {
            "type": [
              "string",
              "null"
            ],
            "format": "date-time",
            "description": "Required when type=assignment"
          },
          "max_score": {
            "type": [
              "string",
              "null"
            ],
            "description": "Optional override; default sum of component max_score"
          }
        }
      },
      "AssessmentDTO": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "id",
          "school_id",
          "version",
          "year_id",
          "term_id",
          "section_id",
          "subject_id",
          "type",
          "max_score",
          "policy_version",
          "state",
          "components"
        ],
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "school_id": {
            "type": "string",
            "format": "uuid"
          },
          "version": {
            "type": "integer",
            "minimum": 1
          },
          "year_id": {
            "type": "string",
            "format": "uuid"
          },
          "term_id": {
            "type": "string",
            "format": "uuid"
          },
          "section_id": {
            "type": "string",
            "format": "uuid"
          },
          "subject_id": {
            "type": "string",
            "format": "uuid"
          },
          "type": {
            "$ref": "#/$defs/AssessmentType"
          },
          "max_score": {
            "$ref": "#/$defs/MarkDecimal"
          },
          "policy_version": {
            "type": "string"
          },
          "state": {
            "$ref": "#/$defs/AssessmentState"
          },
          "components": {
            "type": "array",
            "items": {
              "$ref": "#/$defs/ComponentDTO"
            }
          },
          "due_at": {
            "type": [
              "string",
              "null"
            ],
            "format": "date-time"
          }
        }
      },
...
      "PatchResultRequest": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "expected_version",
          "attempt_id",
          "status"
        ],
        "properties": {
          "expected_version": {
            "type": "integer",
            "minimum": 1
          },
          "attempt_id": {
            "type": "string",
            "format": "uuid"
          },
          "status": {
            "$ref": "#/$defs/ResultWorkflowStatus"
          },
          "marking_outcome": {
            "$ref": "#/$defs/MarkingOutcome"
          },
          "component_scores": {
            "type": "array",
            "items": {
              "$ref": "#/$defs/ComponentScoreInput"
            }
          }
        }
      },
...
      "PatchResultResponse": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "result",
          "version",
          "total"
        ],
        "properties": {
          "result": {
            "$ref": "#/$defs/ResultDTO"
          },
          "version": {
            "type": "integer",
            "minimum": 1
          },
          "total": {
            "type": [
              "string",
              "null"
            ],
            "description": "Computed total; null for absent/exempt"
          }
        }
      },
...
      "BindEvidenceRequest": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "expected_version",
          "pages"
        ],
        "properties": {
          "expected_version": {
            "type": "integer",
            "minimum": 1
          },
          "pages": {
            "type": "array",
            "minItems": 1,
            "items": {
              "$ref": "#/$defs/EvidencePageInput"
            }
          }
        }
      },
...
      "BindEvidenceResponse": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "bindings",
          "version"
        ],
        "properties": {
          "bindings": {
            "type": "array",
            "items": {
              "$ref": "#/$defs/EvidenceRefDTO"
            }
          },
          "version": {
            "type": "integer",
            "minimum": 1
          }
        }
      },
      "ExpectedVersionBody": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "expected_version"
        ],
        "properties": {
          "expected_version": {
            "type": "integer",
            "minimum": 1
          },
          "reason": {
            "type": [
              "string",
              "null"
            ],
            "maxLength": 500
          }
        }
      },
      "PublicationRequest": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "expected_version"
        ],
        "properties": {
          "expected_version": {
            "type": "integer",
            "minimum": 1
          }
        }
      },
      "PublicationResponse": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "publication_id",
          "result_revision_ids",
          "report_job_id"
        ],
        "properties": {
          "publication_id": {
            "type": "string",
            "format": "uuid"
          },
          "result_revision_ids": {
            "type": "array",
            "items": {
              "type": "string",
              "format": "uuid"
            }
          },
          "report_job_id": {
            "type": "string",
            "minLength": 1
          }
        }
      },
...
    },
    "securitySchemes": {
      "session": {
        "type": "apiKey",
        "in": "cookie",
        "name": "sessionid"
      }
    }
  },
...
}